separate controller and procedures created
