+++ db/db-config.js
@@ -2,7 +2,6 @@ const mysql= require("mysql");
 require("dotenv").config();
 
 const dbconnection =()=>{
-  console.log("DB configuration=", process.env.DBUSER,process.env.DBPASSWORD,process.env.DATABASE);
   const dbpool = mysql.createPool({       
     connectionLimit: 20,
     host     : process.env.DBHOST,
