+++ model/userModel.js
@@ -1,31 +1,6 @@
-const bcrypt = require('bcrypt');
-const generateOTP = require('../controller/otp');
-const sendOTPEmail = require('../controller/sendEmail');
-var jwt = require('jsonwebtoken');
-
 const dbconnection = require('../db/db-config');
 const dbpool = dbconnection();
 
-// Function 'checkUserExists' verifies if the user exists
-
-const checkUserExists = async(email, password) => {
-  const userFound = await checkEmailExists(email);
-  if (!userFound) {
-    console.log("Email Mis-match");
-    return null;
-  }
-  // If user is found, compare the password with the hashed password
-  else 
-  {
-    const isHashTrue = await checkHash(password,userFound.password);
-    if(isHashTrue)
-    {
-      return {message:"passMatched", userFound};
-    }
-    else
-      return {message:"nopassMatched"};
-  }
-};
 //Check Email Exists 
 const checkEmailExists = (email) => {
   return new Promise((resolve, reject) => {
@@ -33,6 +8,7 @@ const checkEmailExists = (email) => {
     dbpool.query('call getUserByEmail(?)',[email],(error,results)=>{
       console.log("procedure result",results);
       if (error) {
+        console.log("error in checking mail",error);
         return reject(error);
       }
       // The results structure in case of procedure is [[results[0]],{}]
@@ -44,43 +20,11 @@ const checkEmailExists = (email) => {
     });
   });
 };
-//Compare Hash Function
-async function checkHash(userPassword, hashPassword) {
-  const match = await bcrypt.compare(userPassword, hashPassword);
-  if(match)
-    return true;
-  else
-    return false;
-}
-// Generate Hashed Password
-async function hashPassword(password) {
-  try {
-      const hashedPassword = await bcrypt.hash(password, 10);
-      // Store hashedPassword in your password DB
-      return hashedPassword;
-  } catch (err) {
-      console.error('Error hashing password:', err);
-      throw err; // Or handle the error as appropriate
-  }
-}
 
 // Function to create a new user
-const createUser = async (name,email, password) => {
+const createUser = async (name,email,hashedPassword,token) => {
   try {
-    // Check if the email already exists
-    console.log("Email Recived",email);
-    const emailUser = await checkEmailExists(email);
-    console.log("after email checking",emailUser)
-    if (emailUser) 
-    {
-      console.log("User Exists", emailUser);
-      return { message: "userExists" };
-    }
-    const data = {user:{email:email}};
-    const token = await jwt.sign(data,process.env.JWT_SECRET);
-    console.log("sign up token",token);
-    const hashedPassword = await hashPassword(password);
-    console.log("Hashed Passowrd Generated",hashedPassword);
+    
     // Insert the new user into the database
     //const insertQuery = 'INSERT INTO user (name,email, password,token) VALUES (?,?,?,?)';
       const result = await new Promise((resolve,reject)=>{
@@ -89,21 +33,13 @@ const createUser = async (name,email, password) => {
             return reject(error);
           }
           else{
-            //console.log("add product result=",result);
+            console.log("add user result=",result);
             return resolve(result);
           }
         });
-    
       })
-    //const result = await dbpool.query(insertQuery, [name, email, hashedPassword,token]);
-    console.log("query result of signUp affected Rows=",result.affectedRows);
-    //Generate OTP here
-    const otp = await generateOTP();
-    // send otp to user Email
-    sendOTPEmail(email,otp);
-
     // Return the inserted user
-    return { id: result.insertId,name, email,otp,token };
+    return { name, email,token };
   
   } catch (error) {
     console.log("Some error here in DB",error);
@@ -112,8 +48,7 @@ const createUser = async (name,email, password) => {
 };
 
 // Function to Reset Password
-const resetPassword= async(email,password)=>{
-  const hashedPassword = await hashPassword(password);
+const changePassword= async(email,hashedPassword)=>{
   console.log("Hashed Passowrd Generated",hashedPassword);
   const result = await new Promise((resolve,reject)=>{
     dbpool.query('call updatePassword(?,?)',[email,hashedPassword],(error,result)=>{
@@ -131,8 +66,7 @@ const resetPassword= async(email,password)=>{
   return result;
 }
 module.exports = {
-  checkUserExists,
   createUser,
   checkEmailExists,
-  resetPassword
+  changePassword
 };
